% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anti_addin.R
\name{anti_addin}
\alias{anti_addin}
\title{Create duplicate of addins.dcf file in installed packages.}
\usage{
anti_addin()
}
\value{
nothing
}
\description{
Create duplicates of installed addins.dcf of packages with addins
 found in libpath that rsam will manipulate
}
