% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_keys.R
\name{fetch_addin_keys}
\alias{fetch_addin_keys}
\title{Current Addins shortcuts}
\usage{
fetch_addin_keys()
}
\value{
data.frame
}
\description{
Retrieve current defined Addin shortcuts
}
\seealso{
\code{\link[jsonlite]{fromJSON}}
}
