% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_addins.R
\name{fetch_addins}
\alias{fetch_addins}
\title{Current Addins}
\usage{
fetch_addins(fields = NULL)
}
\arguments{
\item{fields}{character, fields to return from output data.frame, NULL returns all,
Default: NULL}
}
\value{
data.frame containing the columns: c('Package', 'Name', 'Description',
  'Binding', 'Interactive', 'libpath', 'Key', 'Shortcut')
}
\description{
Retrieve current installed Addins
}
