% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_shortcuts.R
\name{fetch_shortcuts}
\alias{fetch_shortcuts}
\title{Current Addins}
\usage{
fetch_shortcuts()
}
\value{
data.frame
}
\description{
Retrieve current Addins and if they have shortcuts they are defined.
}
