% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEYS-data.R
\docType{data}
\name{KEYS}
\alias{KEYS}
\title{Keyboard Validation}
\format{A list with 99 elements of class 'key'}
\usage{
KEYS
}
\description{
List containing keyboard keys
}
\details{
key class has a '.+' method so keys can be appended to each other
}
\examples{
base_key <- KEYS$`left command/window key`
base_key
base_key + KEYS$shift + KEYS$i
base_key + KEYS[2:9]

}
\keyword{datasets}
