% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_addins.R
\name{restore_addins}
\alias{restore_addins}
\title{Restore package addins}
\usage{
restore_addins(package = NULL)
}
\arguments{
\item{package}{character, packages to restore default addins, if NULL all packages will
be restored, Default: NULL}
}
\description{
Restores addins.dcf files of installed package
}
