% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_shorcut.R
\name{rm_shortcut}
\alias{rm_shortcut}
\title{Remove user defined addin keyboard shortcuts}
\usage{
rm_shortcut(fn, verbose = TRUE)
}
\arguments{
\item{fn}{character, binding of addin to remove keyboard shortcut}

\item{verbose}{boolean, print message of actions taken, Default: TRUE}
}
\description{
Command line function to remove user defined keyboard shortcuts for addins
}
\details{
fn expecting the form of the addin binding, ie package::function notation.
}
