% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsam.R
\name{rsam}
\alias{rsam}
\title{rsam user interface}
\usage{
rsam(warn = TRUE)
}
\arguments{
\item{warn}{boolean, show warning popup message to restart IDE, Default: TRUE}
}
\description{
Shiny application to use rsam through UI
}
\seealso{
\code{\link[rhandsontable]{rHandsontableOutput}},\code{\link[rhandsontable]{renderRHandsontable}}
}
