% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_shortcuts-data.R
\docType{data}
\name{rstudio_shortcuts}
\alias{rstudio_shortcuts}
\title{Rstudio Keyboard Shortcuts}
\format{A data frame with 276 rows and 3 variables:
\describe{
  \item{\code{Description}}{character Decription What Shortcut Does}
  \item{\code{Shortcut}}{character Short Definition}
  \item{\code{OS}}{character Operating System}
}}
\usage{
rstudio_shortcuts
}
\description{
Keyboard shortcut definitions scrapped from
\href{https://support.rstudio.com/hc/en-us/articles/200711853-Keyboard-Shortcuts}{RStudio Homesite}
}
\keyword{datasets}
