% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_shortcut.R
\name{set_shortcut}
\alias{set_shortcut}
\title{Set RStudio Addin Keyboard Shortcuts}
\usage{
set_shortcut(fn, shortcut, overide = FALSE, verbose = TRUE)
}
\arguments{
\item{fn}{character, vector of package::function names}

\item{shortcut}{character, vector the same length of fn with corresponding shortcuts}

\item{overide}{boolean, set to TRUE to ignore any existing shortcut, Default: FALSE}

\item{verbose}{boolean, set to TRUE to return a message on actions taken, Default: TRUE}
}
\description{
Helper function that sets Rstudio addin keyboard shortcuts
}
