% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_addin.R
\name{toggle_addin}
\alias{toggle_addin}
\title{Toggle RStudio addins}
\usage{
toggle_addin(key)
}
\arguments{
\item{key}{character, vector containing package::binding objects that uniquely
identify addins.}
}
\description{
Toggle Rstudio addins on and off by passing package::binding
}
\details{
a data.frame containing a summary of installed addins can be fetched through
  \code{\link{fetch_addins}}. Each time a Key is passed into the function the toggle will
  be activated for it.
}
