% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yml.R
\name{create_yml}
\alias{create_yml}
\title{Writes _rsam.yml to home directory}
\usage{
create_yml(overwrite = FALSE)
}
\arguments{
\item{overwrite}{boolean, if TRUE will set output file to defaults, Default: FALSE}
}
\value{
NOTHING
}
\description{
Writes ~/_rsam.yml to bypass rsam onattach/onload prompt menu
}
\details{
Default Settings:

\tabular{lll}{
\strong{field} \tab \strong{description} \tab \strong{setting}\cr
json_no \tab Permission to write ~/.R/rstudio/keybindings/addins.json                     \tab FALSE\cr
dcf_no  \tab Permission to duplicate rstudio/addins.dcf of installed packages with addins \tab FALSE\cr
silent  \tab Permission to run onload/onattach script run with messages to console        \tab  TRUE
}
}
