% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_addins.R
\name{fetch_addins}
\alias{fetch_addins}
\title{Current Addins}
\usage{
fetch_addins(keep_libpath = FALSE)
}
\arguments{
\item{keep_libpath}{boolean, keep package libpath in output object, Default: FALSE}
}
\value{
data.frame
}
\description{
Retrieve current installed Addins
}
