% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raddin.R
\name{raddin}
\alias{raddin}
\title{Raddin Shiny App}
\usage{
raddin(warn = TRUE)
}
\arguments{
\item{warn}{boolean, show warning popup message to restart IDE, Default: TRUE}
}
\description{
UI for Raddin Functionality
}
\seealso{
\code{\link[rhandsontable]{rHandsontableOutput}},\code{\link[rhandsontable]{renderRHandsontable}}
}
